Create generic graphics methods, extract D3D11 code
